fixes some pretty big bugs.

+++ code/_onclick/hud/map_popups.dm
@@ -4,13 +4,16 @@
 /obj/abstract/screen
 	var/assigned_map
 	var/list/screen_info = list()//x,x pix, y, y pix || x,y
+	var/del_on_map_removal = TRUE//this could probably be changed to be a proc, for conditional removal. for now, this works.
 
 /client/proc/clear_map(var/map_to_clear)//not really needed most of the time, as the client's screen list gets reset on relog. any of the buttons are going to get caught by garbage collection anyway. they're effectively qdel'd.
 	if(!map_to_clear|| !(map_to_clear in screen_maps))
 		return FALSE
 	for(var/obj/abstract/screen/x in screen_maps[map_to_clear])
 		screen_maps[map_to_clear] -= x
-		qdel(x)
+		if(x.del_on_map_removal)
+			qdel(x)
+	screen_maps-= map_to_clear
 
 /client/proc/clear_all_maps()
 	for(var/x in screen_maps)
@@ -28,7 +31,7 @@
 	winclone(src,"popupwindow",name)
 	var/list/winparams = new
 	winparams["size"] = "[ratiox]x[ratioy]"
-	winparams["command"] = "handle-popup-close [name]"
+	winparams["on-close"] = "handle-popup-close [name]"
 	winset(src,"[name]",list2params(winparams))
 	winshow(src,"[name]",1)
 
@@ -61,6 +64,7 @@
 	background.layer = -1
 	background.plane = -1
 
+	screen_maps["[popup_name]_map"] += background
 	screen += background
 
 	return newmap
